Fixed a bug with completion of "free" identifiers.
Turn out that when doing autocomple, Vim temporarily erases the
completed part from the buffer lines, so when I'm trying to write to
buffer lines to a temporary file the get written without that part, and
DCD can't read it.

Worked around it by storing that line beforehand(in the first call to
the completion function) and resetting the buffer before writing it
to the temp file.