Fix build on BSD
Added extra version statements for OpenBSD, NetBSD and DragonflyBSD,
as these are not included in the BSD version identifier.

Also removed the line `SHELL:=/usr/bin/env bash' from the top of the
makefile because most BSDs don't include bash in the base system, and
the build doesn't need it anyway.

On OpenBSD, using -inline with dmd causes the compiler to crash
because of insufficient memory (with 8 GB), so I added a couple of
ifneq's to use -inline only if the build is not running on OpenBSD.