tests/extra/tc_ufcs_all_kinds: support gdc
Remove the usage of `-run` which isn't supported by gdc and only pass
`-verrors=0` to dmd and ldc2 since gdc doesn't support the flag and,
by default, all errors are printed.

Change the format in which errors are printed to keep it consistent
across all the three major compilers (a style that they all support is
the gnu style) and make the matching code in generate_tests.d more
readable by using a regex.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>